Update docs: explain how to configure environment
The docs describe what configuration options are available on the environment, but don't spell out how to set those options.  It wasn't clear whether we should set these options as class attributes Environment, or on the instance in use, or by calling a `.configure()` method, etc.

I'm proposing to add a line to the "Environment: Configuration" section explaining that options should be set as attributes on the instantiated environment object.